Test completely done now

+++ src/main/java/de/haw/rnp/messageticker/model/RandomGenerator.java
@@ -45,7 +45,7 @@ public class RandomGenerator {
      * @return sleep time from 1000 - 5000 ms
      */
     public long generateRandomSleepTime() {
-        return (this.generator.nextInt(5000) + 1000);
+        return ((long) (this.generator.nextInt(5000) + 1000));
     }
 
     /**
